requirements.txt
fastapi
uvicorn[standard]
requests
pytest
httpx
python-dotenv
aiosqlitemangum
